fix: avoid possible uncontrolled rollout
previously, just after a completed rollot, we set the status `RolloutCompleted` before resetting the partition of the sts to prevent further rollouts. If between this update of the PhasedRollout to `RolloutCompleted` and the reset of the partition a new version of the sts is set, the controller would go in state "sts partition is now 0, the phased rollot is over", i.e. there would be an uncontrolled rollout.

Fix this by resetting the partition of the sts to prevent further rollouts before setting the PhasedRollout state to `RolloutCompleted`.